Dry up IdFieldMapper implementations (#105586)
The field type is almost the same code across both implementations.
